+++ src/deploy/functions/validate.ts
@@ -89,7 +89,16 @@ export function endpointsAreValid(wantBackend: backend.Backend): void {
   validateTimeoutConfig(endpoints);
   for (const ep of endpoints) {
     validateScheduledTimeout(ep);
-    serviceForEndpoint(ep).validateTrigger(ep, wantBackend);
+    const service = serviceForEndpoint(ep);
+    if (backend.isBlockingTriggered(ep)) {
+      if (service.name === "noop") {
+        throw new FirebaseError(
+          `Unrecognized blocking trigger type: ${ep.blockingTrigger.eventType}. Please update your CLI with ${clc.bold("npm install -g firebase-tools@latest")}.`,
+          { exit: 1 },
+        );
+      }
+    }
+    service.validateTrigger(ep, wantBackend);
   }
 
   // Our SDK doesn't let people articulate this, but it's theoretically possible in the manifest syntax.